+++ CHANGELOG.md
@@ -14,11 +14,14 @@
 * Enchancement: Integration interface better compatibility with Kotlin null-safety
 * Enchancement: Simplify Sentry configuration in Spring integration (#1259)
 * Enchancement: Add Request to the Scope. #1270
+* Fix: Fix SentryTransaction#getStatus NPE (#1273)
+* Enchancement: Optimize SentryTracingFilter when hub is disabled.
+* Fix: SentryTransaction#finish should not clear another transaction from the scope (#1278)
 
 Breaking Changes:
 * Enchancement: SentryExceptionResolver should not send handled errors by default (#1248).
+* Ref: Optimize DuplicateEventDetectionEventProcessor performance (#1247).
 * Enchancement: Add overloads for startTransaction taking op and description (#1244)
-
 
 # 4.1.0